Restructured alg function.
