+++ src/analysis/tax_transfer_funcs/test_tax_transfers/test_alg2.py
@@ -2,9 +2,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from src.analysis.tax_transfer_funcs.benefits import alg2
-from src.analysis.tax_transfer_funcs.benefits import regelberechnung_2011_and_beyond
-from src.analysis.tax_transfer_funcs.benefits import regelberechnung_until_2010
+from src.analysis.tax_transfer_funcs.alg2 import alg2
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_input,
 )
@@ -53,10 +51,10 @@ def test_alg2(year):
     df = load_input(year, file_name, input_cols)
     tb = load_tb(year)
     tb["yr"] = year
-    if year <= 2010:
-        tb["calc_regelsatz"] = regelberechnung_until_2010
-    else:
-        tb["calc_regelsatz"] = regelberechnung_2011_and_beyond
+    # if year <= 2010:
+    #     tb["calc_regelsatz"] = regelberechnung_until_2010
+    # else:
+    #     tb["calc_regelsatz"] = regelberechnung_2011_and_beyond
     calculated = pd.DataFrame(columns=columns)
     for hid in df["hid"].unique():
         calculated = calculated.append(alg2(df[df["hid"] == hid], tb)[columns])
