+++ src/analysis/tax_transfer_funcs/test_tax_transfers/test_kiz.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from src.analysis.tax_transfer_funcs.benefits import kiz
+from src.analysis.tax_transfer_funcs.kiz import kiz
 from src.analysis.tax_transfer_funcs.taxes import kg_eligibility_hours
 from src.analysis.tax_transfer_funcs.taxes import kg_eligibility_wage
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
