+++ src/analysis/tax_transfer_funcs/test_tax_transfers/test_wg.py
@@ -2,7 +2,6 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from src.analysis.tax_transfer_funcs.benefits import wg
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_input,
 )
@@ -12,6 +11,7 @@ from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax impor
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_tb,
 )
+from src.analysis.tax_transfer_funcs.wohngeld import wg
 
 
 input_cols = [
